Implement basic improvements
1) Replace ´randomUniform´ with ´placeholder´.
It should reduce the initial compilation and uploading time for very long waveforms.

2) Add ´factory_reset´ method.
It loads the factory preset to the device.
The method is added to the BaseInstrument class because this method is common to multiples devices: UHFQA, HDAWG.
HDAWG and UHFQA classes call the ´factory_reset´ method from the parent class BaseInsturment.
Update clock rate in ´update_params´ method of parent class ´Sequence´
To me, it makes more sense to update the clock rate in the parent class because clock rate is common to multiple children classes for different sequence types. 

3) Add sequencer command ´play_zero´