Update `init_buffer_indexed` method
Replace `randomUniform` with `placeholder`.
It should reduce the initial compilation and uploading
time for very long waveforms. The target device type is
also added to the arguments so that the granularity
and minimum waveform can be checked depending
on the device type. UHFQA and HDAWG have different
granularities and minimum waveform lengths. A
docstring is also added to the method to improve
documentation. The related test is also updated
according to the changes made to the method.