Add `factory_reset` method
It loads the factory preset to the device.
The method is added to the BaseInstrument class
because this method is common to multiple devices:
UHFQA, HDAWG, UHFLI, MFLI and PQSC.
The instrument classes call the `factory_reset`
method from the parent class BaseInsturment.
Logging facility is also used to allow the user to track
the factory reset event.