Update the enumeration for `TriggerMode`
Replace `EXTERNAL_TRIGGER` with `RECEIVE_TRIGGER`.

`EXTERNAL_TRIGGER` is not completely removed but just deprecated for
the moment. Also add a third trigger mode:
`SEND_AND_RECEIVE_TRIGGER`.

Updated the docstring of `Sequence` class according to the new trigger
enumeration. Ran black to remove the unnecessary spaces at the end of
the lines in the docstring.

The related test is also updated according to the changes made to the
enumeration.