Add helper function `time_to_samples`
This does the some thing as the helper function
`time_to_cycles` with the option `wait_time=False` but it is
simpler and has a more clear naming. Also, we do not need the
`wait_time=True` option of `time_to_cycles` anymore therefore 
I deprecated this helper function now.