Add functions `define_trigger` and `play_trigger`
#### **The following changes are made in this commit:**
##### **1) Add functions `define_trigger` and `play_trigger`:**
* New sequence command is added: `define_trigger`
It generates a waveform of specified length with marker bit 1 set to 1,
analog part set to 0. This is used to send out trigger signals from the
marker output of the AWG. This function is supposed to replace the
`trigger` method, because using marker signals instead of `setTrigger`
command allows us to have more precise timing.
* New sequence command is added: `play_trigger`
It plays the defined trigger from the AWG core output 1. However, this
signal should still be assigned to the desired Marker Output in DIO
settings by selecting `Output 1 Marker 1`.
##### **2) Add tests for `define_trigger` and `play_trigger`:**
##### **3) Deprecate the sequence command: `trigger`:**
This command is outdated. We do not generate triggers using `setTrigger`
anymore because it does not allow us to have precise timing. We should
use marker waveforms to implement more precise triggering.