Update the enumeration for `TriggerMode`
* A new trigger mode is added: `SEND_AND_RECEIVE_TRIGGER`. This trigger
mode will be used when the Master HDAWG is sending out trigger signals
to slave devices, but before that, it waits to receive an external
trigger, which is usually the same 10 MHz clock signal it uses as
external clock. This is necessary to make sure the Master HDAWG
generates the trigger signal always with the same latency with respect
to the external clock it runs from.

* We also replace `EXTERNAL_TRIGGER` with `RECEIVE_TRIGGER` because this
enumeration is more consistent the other enumerations used:
`SEND_TRIGGER` and `SEND_AND_RECEIVE_TRIGGER`. The old enumeration
`EXTERNAL_TRIGGER` is not completely removed but just deprecated for the
moment.

* Also updated the docstring of `Sequence` class according to the new
trigger enumeration.

* The related test is also updated according to the changes made
to the enumeration.