Update `wait_dig_trigger` helper function
#### **The following changes are made in this update:**
##### **1) Update `wait_dig_trigger` helper function:**
The usage of `waitDigTrigger` sequencer command is different for
different devices. For example, in HDAWG it is used as:
`waitDigTrigger(1);`
However, in UHFQA or UHFLI, it is used as:
`waitDigTrigger(1, 1);`
Before this commit, `wait_dig_trigger` helper function was taking this
difference into account, but it was using a confusing indexing with
0,1,2. Also, it was not possible to pass index 2 to the `waitDigTrigger`
command for HDAWG as in:
`waitDigTrigger(2);`
* `wait_dig_trigger` helper function is updated such that it now 
takes the trigger index (1 or 2) and the device type as argument.
Then it generates the correct string using these arguments.
* The usage of the helper function inside the `Sequence` parent class is
 also updated.
##### **2) Update test for `wait_dig_trigger` helper function**