Add new attribute validators
#### **The following changes are made in this commit:**
##### **1) Add new attribute validator `is_greater_equal`:**
This is a generic validator used to check if the value of an attribute
is greater than or equal to a certain minimum value. The attribute can
be either a list or a single element. In case it is a list, it checks if
the smallest element in the list is greater than or equal to the
specified minimum value.
##### **2) Replace validator `is_positive` with `is_greater_equal(0)`:**
Instead of using the validator `is_positive` that can only check if a
value is positive or not, we use the more generic validator
`is_greater_equal` and check if an attribute value is greater than or
equal to 0.
##### **3) Add the `is_greater_equal(0)` validator to attributes<br/>to `wait_cycles` and `dead_cycles`:**
These two attributes did not have a validator to check their value.
Therefore, we add validators to them. We also change the default value
of `wait_cycles` to 28500, corresponding to 95 us. This is difference
between the default values of the period and the dead_cycles.
##### **4) Add new attribute validator `is_smaller_equal`:**
This is a generic validator used to check if the value of an attribute
is smaller than or equal to a certain maximum value. The attribute can
be either a list or a single element. In case it is a list, it checks if
the greatest element in the list is smaller than or equal to the
specified maximum value.
##### **5) Replace `amp_smaller_1` with generic validators:**
Instead of using the validator `amp_smaller_1` that can only check if
the absolute values of all elements in a list of amplitudes is smaller
than or equal to 1.0, we use the more generic validators
`is_greater_equal` and `is_smaller_equal` to check if all elements in
the amplitude list are in the range [-1.0, 1.0].
##### **6) Add new attribute validator `is_multiple`:**
This is a generic validator used to check if the value of an attribute
is multiple of a certain number. It is intended to use with attributes
which need to comply with the granularity specification of an
instrument. The attribute can be either a list or a single element. In
case it is a list, it checks if each element is multiple of the given
factor.