Add new attribute `trigger_samples`
* In the new triggering method, added in one of the previous commits, we
define the trigger signal using the marker bits. The new attribute
`trigger_samples` is used to specify the length of the trigger signal in
number of samples. This attribute should comply with the granularity and
minimum waveform length specifications of HDAWG instrument. Therefore,
necessary validators are added to this attribute.

* A test for attribute `trigger_samples` is also added