Add new attributes related to trigger latency
* New attribute added: `latency_cycles`:
Different instruments have different trigger latencies. In order to
compensate for this, it is necessary for some instruments to wait for
certain number of sequencer cycles after receiving the trigger. For
instance, HDAWG needs wait for 27 sequencer cycles after receiving the
trigger because its trigger latency is lower than UHFQA. This way, a
waveform from HDAWG can be aligned with a waveform from UHFQA. In this
case, UHFQA does not need to wait any extra cycles after receiving the
trigger. `latency_cycles` specifies how many cycles an instrument should
wait after receiving the trigger. By default, it is 27 for HDAWG and 0
for UHFQA/UHFLI. This attribute is not accessible directly to the user.

* New attribute added: `latency_adjustment`:
Since `latency_cycles` is not accessible to the user, we introduce
another attribute to allow the user to adjust the number of cycles an
instrument waits for after receiving the trigger. Using the
`latency_adjustment` attribute, the user can increase `latency_cycles`
from its default value. It is set to 0 by default. A test is added to
check if this attribute can be accessed and changed by the user.