Add attributes `wait_samples` and `dead_samples`
* New attribute added:`wait_samples`:
Wait time is the portion of the period before t0. Therefore, it is
calculated as: wait time = period - dead_time + trigger_delay
The attribute `wait_samples` is the wait time calculated as above and
converted to units of number of samples. It has a default value
corresponding to 95 us on an HDAWG instrument.

* New attribute added:`dead_samples`:
Dead time is the portion of the period after t0. The attribute
`dead_samples` is the dead time converted to units of number of samples.
It has a default value corresponding to 5 us on an HDAWG instrument.