Add new attributes related to trigger commands
These attributes are set depending on the trigger mode and trigger
latency compensation

* New attribute added `trigger_cmd_latency`:
This attribute is set to a `wait` command with the argument
`latency_cycles` passed to it if trigger `latency_cycles` is not 0, so
that the instrument waits for `latency_cycles` in order to compensate
for the trigger latency differences in different instruments. An inline
comment is also added to inform the user about this.

* New attribute added `trigger_cmd_define`:
If the trigger mode is `SEND_AND_RECEIVE_TRIGGER` or `SEND_TRIGGER`, it
is set to the `define_trigger` sequencer command with the argument
`trigger_samples` passed to it.

* New attribute added `trigger_cmd_send`:
If the trigger mode is `SEND_AND_RECEIVE_TRIGGER`, this attribute is set
to combination of sequencer commands `wait_dig_trigger` and
`play_trigger`. In this trigger mode, the master instrument waits for
an external clock before sending out the trigger signal. An inline
comment is added to the sequencer to inform the user about this. After
receiving the external clock, the trigger signal is sent out. If the
trigger mode is `SEND_TRIGGER`, this attribute is set to only
`play_trigger` sequencer command.

* New attribute added `trigger_cmd_wait`:
If the trigger mode is `SEND_AND_RECEIVE_TRIGGER` or `SEND_TRIGGER`,
this attribute is set to `wait_dig_trigger` sequencer command, so that
the instrument waits for self triggering. An inline comment is added to
the sequencer to inform the user about this. Self triggering is
necessary for master devices in order to ensure proper alignment of
waveforms with slave devices. If the trigger mode is `RECEIVE_TRIGGER`,
this attribute is set to `wait_dig_trigger` sequencer command, so that
the instrument waits for an external trigger. An inline comment is added
to the sequencer to inform the user about this.

All attributes are set to comment line when they are not used in a
certain trigger mode.