Add `readout_cmd_trigger`, `osc_cmd_reset` attribs
* New attribute added `readout_cmd_trigger`:
This attribute depends on the device type. It is used to trigger the
quantum analyzer of an instrument. Since only UHFQA has quantum
analyzer, it is set to comment line in other devices. For UHFQA it is
set to `readout_trigger` sequencer command.

* New attribute added `osc_cmd_reset`:
This attribute is used for setting the oscillator phase to 0 if the
`reset_phase` attribute of `Sequence` parent class is set to `True`.
The sequencer command `reset_osc_phase` is used for this purpose. If
`reset_phase` attribute is set to `False, `osc_cmd_reset` is set to
 comment line.

* The docstring is updated to remove the part that states `reset_phase`
option is only available for HDAWG. This option works also for UHFQA
and UHFLI.

* A test is added to check if the user can access and change the
attribute `reset_phase`.