Update sequence commands
* Update helper function `wait`:
This helper function is updated to add a description and return "//" if
waiting time given as input is 0.
* Update helper function `readout_trigger`:
This helper function is updated to use the new QA starting method.
* Update helper function `repeat`:
This helper function is updated to remove unnecessary new line
characters. 
* Update helper function `close_bracket`:
This helper function is updated to remove unnecessary new line
characters in front of the closing bracket '}' . 
* Add new helper function `assign_wave_index`:
This helper function is used to assign index number to the labeled
waveforms.
* Add new helper function `inline_comment`
This command allows us to insert inline comments to the sequencer.
* Add new helper function `space`
This command allows us to insert space character to the sequencer.
* Add new helper function `tab`
This command allows us to insert tab character to the sequencer.
* Add new helper function `header_info`:
This helper function is used to display the sequence type, trigger mode
and alignment information at the top of the sequencer program. The
outdated function `header_comment` was only displaying the sequence type
and it is deprecated.
* Related tests are created/updated for these sequence commands.