Update `write_sequence` method
* The sequencer program is initiated in this method. The header
information of the sequencer program is now written in the
`write_sequence` method of the parent class `Sequence`. Sequence type is
temporarily selected as `None` here. It will be overwritten by the
children classes depending on the actual sequence type.

* `get_sequence` test is also updated. When the sequence type is `None`,
the sequence program is not equal to `None` anymore. It is initiated
with the header information and the sequence type is state as `None` in
this header information, such as:
```
// Zurich Instruments sequencer program
// sequence type:              None
// trigger mode:               Send Trigger
// alignment:                  End with Trigger
// automatically generated:    30/04/2021 @16:17
```