Update ´TriggerSequence´ class
#### **The following changes are made to `TriggerSequence` class:**

##### **1) Docstring is updated:**
The docstring is changed to reflect the new triggering method.
##### **2) Header information is updated:**
The header information inserted to the top of the program is updated. It
now includes the trigger mode and alignment as well as the sequence
type. The header information is initiated as a string in the parent
class `Sequence` and written into the `sequence` attribute. The
`TriggerSequence` class calls the `write_sequence` method of the parent
class to get the header information. It then replaces the string 'None'
with the correct sequence type.
##### **3) New commands are integrated into the sequence:**
The commands `trigger_cmd_define` and `trigger_cmd_send` are now being
used to generate the sequence. Outdated commands are not being used
##### **4) Improved explanation of the code with inline comments:**
We use `inline_comment` helper function to insert explanatory comments
to the sequencer program.
##### **5) Added support for trigger mode `SEND_AND_RECEIVE_TRIGGER`:**
The `Trigger` sequence type now supports the new trigger mode
`SEND_AND_RECEIVE_TRIGGER` in addition to `SEND_TRIGGER`. If the
selected trigger mode is neither of these two, then the trigger mode is
automatically set to `SEND_TRIGGER`. The docstring is updated to inform
the user about this. Also a logger message is displayed to report to the
user that their trigger mode selection is overwritten. The related tests
are also updated to add support for `SEND_AND_RECEIVE_TRIGGER`.