Update ´PulsedSpectroscopySequence´ class
#### **The following changes are made to this class:**

##### **1) Docstring is updated:**
The docstring is changed reflect the changes made to this class
and follow the line length limit of 72.
##### **2) Header information is updated:**
The header information inserted to the top of the program is updated. It
now includes the trigger mode and alignment as well as the sequence
type. The header information is initiated as a string in the parent
class `Sequence` and written into the `sequence` attribute. The
`PulsedSpectroscopySequence` class calls the `write_sequence` method of
the parent class to get the header information. It then replaces the
string 'None' with the correct sequence type.
##### **3) `pulse_samples` attribute added:**
This attribute is specific to the child class
`PulsedSpectroscopySequence`. It is used to convert the pulse length to
the number of samples.
##### **4) `wait_samples_updated` and ` dead_samples_updated` added:**
These attributes are used to update the number of samples to wait before
and after playing each waveform. The pulse length is not taken into
account while the calculations are done in the parent class `Sequence`.
Therefore, we update these attributes in the child class
`PulsedSpectroscopySequence`. Note that how these attributes  are
updated depends on the alignment option. The update is done inside
`update_params` method after calling `super().update_params()` and
making the necessary updates to `pulse_samples` first. Outdated
variables are not being used anymore: `temp`, `wait_cycles`,
`dead_cycles`.
##### **5) `reset_phase` attribute hardcoded to be `True`:**
This sequence type requires the oscillator phase to be reset before
playing the pulse. Therefore, the user preference is ignored and
`reset_phase` attribute is set to 0 inside `update_params` method. The
related test is also updated to to respond this change.
##### **6) New commands are integrated into the sequence:**
The commands trigger_cmd_wait`, `trigger_cmd_latency`, `osc_cmd_reset`,
`readout_cmd_trigger` are now being used to generate the sequence.
Outdated commands are not being used.
##### **7) Improved explanation of the code with inline comments:**
We use `inline_comment` helper function to insert explanatory comments
to the sequencer program.