Embed necessary settings for pulsed spectroscopy
#### **The following changes are made to UHFQA driver:**
##### **1) Implement ´_qa_delay_default´ method:**
This method is added to the `UHFQA` class because it is specific to this
instrument. It is used to automatically determine the default value of
quantum analyzer delay depending on enabling of deskew matrix. The
method is only for internal use.
##### **2) Implement ´qa_delay´ method:**
This method is added to the `UHFQA` class because it is specific to this
instrument. It is added to allow the user to make adjustments to quantum
analyzer delay. If no argument is passed to it, it returns the current
value of the QA delay adjustment. A test is added for this method as
well.
##### **3) Add `_qa_delay_user` attribute:**
This attribute holds the current value of the quantum analyzer delay
adjustment. By default it is set to 0. It is not directly accessible to
the user but the user can change it using the ´qa_delay´ method.
##### **4) Docstring edited:**
Docstring is updated to respect the line length limit of 72. Also one
line of example code added to show how to use the `qa_delay` method.
##### **5) Embed the settings for pulsed spectroscopy:**
`_apply_pulsed_settings` of the device-specific AWG Core for UHFQA is
updated to embed the necessary settings for pulsed spectroscopy into the
UHFQA driver to increase automation.