Improve PQSC controller
#### **The following changes are made in this update:**
##### **1) `connect_device` method is added:**
This method is used to connect the PQSC to the data server.
##### **2) `arm` method is added:**
This method prepares the PQSC to send out trigger signals. It first,
disables the PQSC if it is running. Then it configures the number of
triggers and the time between repeated trigger, using the optional
arguments `repetitions` and `holdoff`, respectively. Then it clears the
register bank.
##### **3) `run` and `stop` methods are added:**
These methods are used to enable and disable the PQSC.
##### **4) New parameters added to PQSC:**
`ref_clock` is used to set and get the reference clock source for the
PQSC. `ref_clock_actual` is used to get the actual reference clock
source the PQSC is configured to at that moment. Note that the actual
reference clock source can be different then the intended clock source
for a short time since it takes some time for the PQSC to lock to the
clock source. `ref_clock_status` is used to get the locking status of
the reference clock. New parsers are implemented for these parameters.
`progress` is used to get the percentage of repeated triggers sent out
so far.
##### **5) `check_ref_clock` method is added:**
This method is used to check if the reference clock is locked or not. If
the argument `blocking` is set to true, it waits until the locking is
successfully completed. The maximum amount of time to wait is specified
by the `timeout` parameter.
##### **6) HDAWG controller is updated:**
`ref_clock` and `ref_clock_status` parameters are added, similar to PQSC
driver. `enable_manual_mode` method is added to reset the HDAWG to
default settings to disable QCCS mode.
##### **7) UHFQA controller is updated:**
`ref_clock` parameter is added, similar to PQSC and HDAWG drivers.
`enable_manual_mode` method is added, similar to HDAWG driver.
##### **8) AWG node existence check added to `connection.py`:**
Inside `connect_device` method, the AWG module is updated only if the
device has AWG functionality.
##### **9) Updated tests for PQSC, HDAWG, UHFQA and parsers