Update `HDAWG` and `UHFQA` parameters
#### **The following changes are made in this update:**

##### **1) Add `_check_node_exists` method to `BaseInstrument` class:**
This method is only for internal use. It is used to check if the
specified node exists in the device.
##### **2) Add `_get_node_dict` method to `BaseInstrument` class:**
This method is only for internal use. It is used to retrieve the
dictionary associated with the specified node.
##### **3) Add `_init_params` method to `BaseInstrument` class:**
This method is only for internal use. It is called inside
`connect_device` method to initialize the device parameters associated
with the nodes of the device.
##### **3) Add `_init_awg_params` method to `AWGCore` class:**
This method is only for internal use. It is called inside
`_init_awg_cores` method of the `HDAWG` and `UHFQA` classes to
initialize the device AWG parameters associated with the nodes of the
device AWGs.
##### **4) Update `_init_` method of `HDAWG` class:**
AWG Core class is not called in this method anymore. The `_awgs`
attribute is initiated as an empty list instead. Two parameters added
and initiated as `None`: `ref_clock` and `ref_clock_status`
##### **5) Add `_init_params` method to `HDAWG` class:**
This method is only for internal use and it overwrites the
`_init_params` method of `BaseInstrument`. It expands the parameters
initiated in `_init_` method. `ref_clock` is used to set and get the
reference clock source for the HDAWG. `ref_clock_status` is used to get
the locking status of the reference clock. The node dictionaries
associated with these parameters are extracted from the device using
`_get_node_dict` method.
##### **6) Add `_init_awg_params` method to HDAWG `AWG` child class:**
This method is only for internal use. It is similar to `_init_params`
method of `HDAWG` class but is is a method of `AWG` child class of
`HDAWG` class. It overwrites th `_init_awg_params` method of `AWGCore`
class. It expands the parameters initiated in `_init_` method of `AWG`
child class. The node dictionaries associated with these parameters are
extracted from the device using `_get_node_dict` method.
##### **7) Add `_init_awg_cores` method to `HDAWG` class:**
This method is only for internal use and it is called inside the
`connect_device` method of the `HDAWG` class. It calls the AWG Core
class to initialize device AWG Cores and assigns them to the list
`_awgs`. Then AWG modules are initiated for each core. Then the AWG
parameters are initiated for each core using `_init_awg_params` method.
##### **8) Update `_init_` method of `UHFQA` class:**
`AWG` Core class and `ReadoutChannel` class are not called in this
method anymore. The `_awg` attribute is initiated as `None.` The
`_channels` attribute is initiated as an empty list. Other parameters
are initiated as `None`. A new parameter added: `ref_clock`. It is used
to set and get the reference clock source for the UHFQA.
##### **9) Add `_init_params` method to `UHFQA` class:**
This method is only for internal use and it overwrites the
`_init_params` method of `BaseInstrument`. It expands the parameters
initiated in `_init_` method. The node dictionaries
associated with these parameters are extracted from the device using
`_get_node_dict` method.
##### **10) Add `_init_awg_params` method to UHFQA `AWG` child class:**
This method is only for internal use. It is similar to `_init_params`
method of `UHFQA` class but is is a method of `AWG` child class of
`UHFQA` class. It overwrites th `_init_awg_params` method of `AWGCore`
class. It expands the parameters initiated in `_init_` method of `AWG`
child class. The node dictionaries associated with these parameters are
extracted from the device using `_get_node_dict` method.
##### **11) Add `_init_channel_params` method to `ReadoutChannel`:**
This method is only for internal use. It is similar to `_init_params`
method of `UHFQA` class but is is a method of `ReadoutChannel` child
class of `UHFQA` class. It expands the parameters initiated in `_init_`
method of `ReadoutChannel` child class. The node dictionaries associated
with these parameters are extracted from the device using
`_get_node_dict` method.
##### **12) Add `_init_awg_cores` method to `UHFQA` class:**
This method is only for internal use and it is called inside the
`connect_device` method of the `UHFQA` class. It calls the AWG Core
class to initialize device AWG Core and assigns it to the attribute
`_awg`. Then AWG module and AWG parameters are initiated using `_setup`
and `_init_awg_params` methods, respectively.
##### **13) Add `_init_readout_channels` method to `UHFQA` class:**
This method is only for internal use and it is called inside the
`connect_device` method of the `UHFQA` class. It calls the
`ReadoutChannel` class to initialize device readout channels and assigns
them to the list `_channels`. Then readout channel parameters are
initiated using `_init_channel_params`method for each channel.
##### **14) New parsers added for new parameters.**
##### **15) Tests for drivers are updated.**