Update parameters
Update parameters

#### **The following changes are made in this update:**
##### **1) `Parameter` class in `node_tree.py` is updated:**
* A new input is added: `auto_mapping`. This setting specifies whether
an automatic mapping should be constructed from the allowed options
information obtained from the device. A private function
`_construct_auto_mapping` is added to construct the mapping. Note that,
automatic mapping construction only happens if no manual mapping is
specified in the parameter settings.
* `_getter` and `_setter` functions are updated to work with mappings
which have several values assigned to the same key. This can be the case
when mapping is automatically generated from the options of the node.
New helper functions `_invert_mapping` and `_flatten_mapping_values` are
added for `_getter` and `_setter`. Also the `_setter` function is
updated to deal with list of set_parser functions as well as single
set_parser function.
* `_reformat_options_dict` method is added to insert new line characters
between each allowed value in options so that it the options are
displayed in a more human readable format in the docstring.
##### **2) `Parse` class in `parsers.py` is updated:**
* Some of the parsers used for mapping are removed because the mapping 
is now automatically constructed from node options. The parsers 
`greater0`, `amp1` and `abs90` are removed since they were too specific.
 More general parsers are added: `greater`, `greater_equal`, `smaller`,
`smaller_equal`, `multiple_of`. The parsers `uhfqa_time2samples` and
`uhfqa_samples2time` are updated with the new added parsers.
* Warning loggers are added to warn the user when the parsers round a 
value automatically since the user input does not match the value limit 
or multiplicity requirements. No exceptions are raised in this case and 
the program continues running with the rounded values, similar to have 
LabOne GUI behaves.
##### **3) UHFQA driver updates:**
The parameter `integration_time` is updated to correct its description.
A new parameter `integration_length` is added. Docstring is updated.
##### **4) HDAWG driver updates:**
Docstring is updated.
##### **5) PQSC driver updates:**
The content of `check_ref_clock` function is moved to `BaseInstrument`
class and made private. This private function is called inside the
device specific drivers if the device supports it. Currently only PQSC
and SHFQA support this function. Docstring is updated.
##### **6) Tests are updated**