Update PQSC, HDAWG and UHFQA drivers
#### **The following changes are made in this update:**
##### **1) An interface for command table is added to HDAWG driver:**
This interface allows users to upload their command table to the device
##### **2) `single` parameter is added to HDAWG and UHFQA AWG:**
This parameter specifies if rerun is enabled or not in the AWG.
##### **3) `check_zsync_connection` method is added to PQSC driver:**
This method checks if the ZSync connection on the given port is
successful or not.
##### **4) `factory_reset` method of UHFQA is updated:**
This method now enables the single mode of AWG and reset the QA delay to
its default value. Setting these manually is necessary because
these settings are not done automatically at power on or after loading
the factory preset to the device.
##### **5) `int_weights_envelope` method is added to UHFQA:**
This method is used to take a list of envelope values to multiply with
the integration weights. `_set_int_weights` and `_demod_weights` are
updated to calculate the integration weights based on the
`int_weights_envelope`.
##### **6) `compile` method of AWG is updated :**
This method is updated to add `logging.warning` and `logging.error`
depending on the compiler status. These logger messages guides the user
better for solving the issues with the seqc program. Also, it is now
possible to upload waveforms in `Custom` sequence type.