Fix bugs and improvements
#### **The following changes are made in this update:**
##### **1) UHFQA updates and bug fixes:**
* `_init_settings` method is updated. The AWG object and parameters
should not be used inside this method because when this method is
called, the AWG and the parameters are not initiated yet and they are
set to `None`. The node addresses should be used instead.
##### **2) PQSC updates and bug fixes:**
* `is_running` attribute added.
* `enable`, `repetitions` and `holdoff` parameters are added.
* `run, `stop` and `arm` methods are updated.
Fix bugs and improvements

#### **The following changes are made in this update:**
##### **1) UHFQA updates and bug fixes:**
* `_init_settings` method is updated. The AWG object and parameters
should not be used inside this method because when this method is
called, the AWG and the parameters are not initiated yet and they are
set to `None`. The node addresses should be used instead.
##### **2) PQSC updates and bug fixes:**
* `is_running` attribute added.
* `enable`, `repetitions` and `holdoff` parameters are added.
* `run, `stop` and `arm` methods are updated.