Update `BaseInstrument` class
#### **The following changes are made in this update:**
##### **1) New parameters added to `BaseInstrument`:**
`data_server_version`, `firmware_version` and `fpga_version` are
parameters common to all devices. Therefore, they are added to the
`BaseInstrument` class.
##### **2) `_get_node_dict` method is updated:**
`_command_to_node` method from `DeviceConnection` class is called to
 add the device serial to the node string if it does not start with
 '/zi/'. This check is necessary because the `data_server_version`
 parameter node starts with '/zi/' and no device serial should be added
 to this node string.
##### **3) `_check_node_exists` method is updated:**
This method now takes as argument the full node address with the device
serial is added.
##### **4) `_init_params` method of device specific drivers updated:**
The device specific drivers of HDAWG, should first call the method from
the `BaseDriver` to initialize the parameters common to all drivers.
Then they initialize the parameters specific to that instrument.
##### **5) `version_parser` added to `Parse` class:**
This method is used to convert the obtained data version string to
`{year}.{month}.{build}` format
##### **6) `_options` attribute initialized as `None`**
##### **7) `test_baseInstrument` tests are improved:**
* Test serial number is changed to a more realistic number.
* `list_nodes` method in `ConnectionMock` is updated to be more
realistic.
* Tests added for `_get_streamingnodes` and `options` attribute
* Test for `_get_node_dict` method updated.
##### **8) `test_version_parser` added to `test_parsers`**