Update `UHFQA` class
#### **The following changes are made in this update:**
##### **1) `_init_settings` method is updated:**
* The AWG object and its parameters should not be used inside this
method because when this method is called, the AWG core and its
parameters are not initiated yet and they are set to `None`. The node
addresses should be used instead to set the initial settings.
* The `qa_delay` adjustment is set to 0 as initial setting.
##### **2) `single` attribute is initialized as `None`**
##### **3) `__repr__` method of `AWGCore` class is updated:**
* The device name is removed from the string representation `__repr__`
of the AWG object.
##### **4) `int_weights_envelope` method is updated:**
If the envelope is given as a list, the length of the list is compared
with the integration length and a `ToolkitError` is raised if they do
not match.
##### **5) Documentation improved:**
* A more realistic device serial is used in the docstring.
* Docstring of `enable_qccs_mode` is updated.
* Docstring is added to `enable_manual_mode`
* The docstring of `AWG` class is improved.
* Typos corrected in the comments and docstrings.
* Type of attributes changed from Parameter to zhinst.toolkit.control
.node_tree.Parameter since the former one does not link to the Parameter
 class in the documentation.
##### **6) `test_uhfqa` tests are improved:**
* Test serial number is changed to a more realistic number.
* `test_init_uhfqa` test is updated to check if
`integration_length`, `averaging_mode` and `ref_clock` are initialized
as `None` and `_qa_delay_user` is initialized as 0.
* `test_qa_delay` is updated to test `_qa_delay_default`
* `test_init_uhfqa_awg` is updated. It checks if `single` is initiated
as `None`. `update_readout_params` is removed from this test. `compile`
is called after setting the sequence type.
* test_update_readout_params test is added to test
`update_readout_params`. When the sequence type is
`SequenceType.READOUT`, the readout channel is enabled and calling
`update_readout_params` does not raise an Exception. For other sequence
types, `update_readout_params` should raise an  Exception.
* `test_init_readout_channel` is updated to check the `_parent`,
`_enabled`, `_readout_frequency`, `_readout_amplitude`,
`_int_weights_envelope` and `_phase_shift` attributes.
* `test_int_weights_envelope` added to test `int_weights_envelope`
method.