Update `ZIConnection` and `DeviceConnection`
#### **The following changes are made in this update:**
##### **1) `setVector` method of is renamed to `set_vector`:**
The function names should be lowercase.
##### **2) Attributes of `ZIConnection` updated:**
* `_awg` attribute is removed since it was unused.
* Attributes `_awg_module`, `_daq_module` and `_sweeper_module` are
initialized as `None`
* Public attribute `daq` is added as it is described in the docstring.
* The modules calling the protected attribute `_daq` are updated
to call the public attribute `daq`. This public attribute
performs a connection check before returning the `_daq`
attribute.
##### **3) `DeviceConnection` class is updated:**
* Attributes `normalized_serial` and `discovery` are initialized and
used inside the class as protected attributes.
* Public attributes `connection`, `device`, `normalized_serial`,
`discovery`, `is_established` and `_is_connected` are added. The
attributes  `connection` and `device` are already described in the
docstring. A description is added for  `normalized_serial`, `discovery`,
`is_established` and `_is_connected` to the docstring.
* The modules calling the protected attribute `_connection` are updated
to call the public attribute `connection`. This public attribute
performs a connection check before returning the `_connection`
attribute.
* `connect_device` method is updated to check if a connection to data
server is established before trying to connect the device to the data
server.
##### **4) `BaseInstrument` class updated:**
* Description for `options` attribute is added to docstring
* Unused attribute `_normalized_serial` is removed
* `connect_device` method is updated. The connection check is removed
since it is done inside the method `_controller.connect_device`
* `_check_connected` method of is updated. If the device is not
connected to the data server, a `ToolkitConnectionError` is raised
instead of `ToolkitError`.
* `is_connected` attribute is updated to return the flag showing if the
device is connected to the data server as it is described in the
docstring. Previously, it was returning a flag showing if a connection
to data server is established.
* `serial` attribute is updated to perform a check if the device is
connected to the data server.
##### **5) Following tests are updated:**
* `test_baseInstrument.py`
* `test_connection.py`
* `test_shfqa.py`